+++ lib/src/pcanvas_impl_html.dart
@@ -1,5 +1,6 @@
 import 'pcanvas_base.dart';
 import 'pcanvas_html.dart';
 
-PCanvas createPCanvasImpl(int width, int height, PCanvasPainter painter) =>
-    PCanvasHTML(width, height, painter);
+PCanvas createPCanvasImpl(int width, int height, PCanvasPainter painter,
+        {PCanvasPixels? initialPixels}) =>
+    PCanvasHTML(width, height, painter, initialPixels: initialPixels);